plugin adjustment.

+++ own-warehouse-boss/bin/catalina.sh
@@ -12,25 +12,27 @@ cd ..
 case "$1" in
 
     start)
-        nohup java -Xms256m -Xmx512m -jar $JAR_NAME --server.port=8080 --spring.config.location=file:conf/db.properties >/dev/null 2>&1 &
+        nohup java -Xms256m -Xmx512m -jar $JAR_NAME --server.port=8080 --spring.config.location=file:conf/db.properties >nohup.log 2>&1 &
         echo $! > $PID
         echo "=== start $SERVICE_NAME"
         ;;
 
     stop)
-        kill `cat $PID`
-        rm -rf $PID
-        echo "=== stop $SERVICE_NAME"
+        if [ -f "$PID" ]; then
+            kill `cat $PID`
+            rm -rf $PID
+            echo "=== stop $SERVICE_NAME"
 
-        sleep 5
-        ##
-        P_ID=`ps -ef | grep -w "$SERVICE_NAME" | grep -v "grep" | awk '{print $2}'`
-        if [ "$P_ID" == "" ]; then
-            echo "=== $SERVICE_NAME process not exists or stop success"
-        else
-            echo "=== $SERVICE_NAME process pid is:$P_ID"
-            echo "=== begin kill $SERVICE_NAME process, pid is:$P_ID"
-            kill -9 $P_ID
+            sleep 5
+            ##
+            P_ID=`ps -ef | grep -w "$SERVICE_NAME" | grep -v "grep" | awk '{print $2}'`
+            if [ "$P_ID" == "" ]; then
+                echo "=== $SERVICE_NAME process not exists or stop success"
+            else
+                echo "=== $SERVICE_NAME process pid is:$P_ID"
+                echo "=== begin kill $SERVICE_NAME process, pid is:$P_ID"
+                kill -9 $P_ID
+            fi
         fi
         ;;
